ENGR00233809: Ethernet:fix two NFS issues as filesystem
Introduce property "ro.nfs.mode" default no to control two cases which will
broken NFS boot as filesystem.
1.For android4.2,drop_privilege in installd will let it have no permission
to install packages because of security policy's limitation in NFSV3.
2.Another case is suspend/resume in NFS mount.it will report "nfs: server
10.192.*.* not responding, still trying",This is caused by clearing ethernet
interface addresses. so the netd will not pass ethernet's link up message to
framework. In this case, no need to clear it for we keep on connecting to the
same network and no need to do dhcp again.

Signed-off-by: Jianzheng Zhou <jianzheng.zhou@freescale.com>

Conflicts:
	cmds/installd/installd.c